Consolidate metadata
Add support for --build-system in elliott

revert unrelated runtime change

tweak imports

Import fixes

Fix imports

Add pylintrc and make task

Move pylintrc

pylint: do recursive search for imports

Fix elliott functional test dir
