+++ doozer/tests/test_metadata.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock, Mock, patch
 from artcommonlib.model import Model
 from doozerlib.image import ImageMetadata
 from doozerlib.metadata import Metadata, CgitAtomFeedEntry, RebuildHintCode
-from doozerlib.brew import BuildStates
+from artcommonlib.brew import BuildStates
 
 
 class TestMetadata(TestCase):
