+++ elliott/elliottlib/bzutil.py
@@ -1290,7 +1290,8 @@ async def approximate_cutoff_timestamp(basis_event: int, koji_api: ClientSession
     """ Calculate an approximate sweep cutoff timestamp from the given basis event
     """
     basis_timestamp = koji_api.getEvent(basis_event)["ts"]
-    builds: List[Dict] = await asyncio.gather(*[exectools.to_thread(meta.get_latest_build, default=None, complete_before_event=basis_event, honor_is=False) for meta in metas])
+    tasks = [meta.get_latest_build(default=None, complete_before_event=basis_event, honor_is=False) for meta in metas]
+    builds: List[Dict] = await asyncio.gather(*tasks)
     nvrs = [b["nvr"] for b in builds if b]
     rebase_timestamp_strings = filter(None, [isolate_timestamp_in_release(nvr) for nvr in nvrs])  # the timestamp in the release field of NVR is the approximate rebase time
     # convert to UNIX timestamps
