+++ elliott/elliottlib/cli/find_unconsumed_rpms.py
@@ -108,7 +108,7 @@ class FindUnconsumedRpms:
         # Get image builds for the assembly
         image_metas: List[ImageMetadata] = [image for image in self._runtime.image_metas() if not image.base_only and image.is_release]
         logger.info("Fetching Brew builds for %s component(s)...", len(image_metas))
-        brew_builds: List[Dict] = await asyncio.gather(*[exectools.to_thread(image.get_latest_build) for image in image_metas])
+        brew_builds: List[Dict] = await asyncio.gather(*[image.get_latest_build() for image in image_metas])
 
         logger.info("Retrieve RPMs in %s image build(s)...", len(brew_builds))
         build_archives = FindUnconsumedRpms._list_archives_by_builds([b["id"] for b in brew_builds], "image", koji_api)
