+++ elliott/tests/test_metadata.py
@@ -6,7 +6,7 @@ import datetime
 from unittest.mock import MagicMock, Mock
 
 from elliottlib.metadata import Metadata
-from elliottlib.brew import BuildStates
+from artcommonlib.brew import BuildStates
 
 
 class TestMetadata(unittest.TestCase):
@@ -117,14 +117,14 @@ class TestMetadata(unittest.TestCase):
 
         # If listBuilds returns nothing, no build should be returned
         builds = []
-        self.assertIsNone(meta.get_latest_build(default=None))
+        self.assertIsNone(meta.get_latest_brew_build(default=None))
 
         # If listBuilds returns a build from an assembly that is not ours
         # get_latest_builds should not return it.
         builds = [
             self.build_record(now, assembly='not_ours', el_target=8)
         ]
-        self.assertIsNone(meta.get_latest_build(default=None))
+        self.assertIsNone(meta.get_latest_brew_build(default=None))
 
         # If there is a build from the 'stream' assembly, it should be
         # returned.
@@ -132,7 +132,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now, assembly='stream', el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])
 
         # Filtering should prefer images which match our tag's RHEL version, bug
         # if there is no match for our elX, at the very least, it should filter out
@@ -142,13 +142,13 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='stream', el_target=8),
             self.build_record(now, assembly='stream', el_target=9)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])
 
         # If there is a build for our assembly, it should be returned
         builds = [
             self.build_record(now, assembly=runtime.assembly, el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[0])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[0])
 
         # If there is a build for our assembly and stream, our assembly
         # should be preferred even if stream is more recent.
@@ -157,7 +157,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now, assembly=runtime.assembly, el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[2])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[2])
 
         # The most recent assembly build should be preferred.
         builds = [
@@ -166,7 +166,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now, assembly=runtime.assembly, el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[3])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[3])
 
         # Make sure that just matching the prefix of an assembly is not sufficient.
         builds = [
@@ -175,7 +175,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now, assembly=f'{runtime.assembly}b', el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])
 
         # el7 should not match.
         builds = [
@@ -184,14 +184,14 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now, assembly=f'{runtime.assembly}', el_target=7)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])
 
         # By default, we should only be finding COMPLETE builds
         builds = [
             self.build_record(now - datetime.timedelta(hours=5), assembly='stream', el_target=8, build_state=BuildStates.COMPLETE),
             self.build_record(now, assembly='stream', el_target=8, build_state=BuildStates.FAILED),
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[0])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[0])
 
         # By default, we should only be finding COMPLETE builds
         builds = [
@@ -199,7 +199,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly=None, el_target=8, build_state=BuildStates.FAILED),
             self.build_record(now, assembly=None, el_target=8, build_state=BuildStates.COMPLETE),
         ]
-        self.assertEqual(meta.get_latest_build(default=None, assembly=''), builds[2])
+        self.assertEqual(meta.get_latest_brew_build(default=None, assembly=''), builds[2])
 
         # Check whether extra pattern matching works
         builds = [
@@ -209,7 +209,7 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', el_target=8),
             self.build_record(now - datetime.timedelta(hours=8), assembly=f'{runtime.assembly}', el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None, extra_pattern='*.g1234567.*'), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None, extra_pattern='*.g1234567.*'), builds[1])
 
     def test_get_latest_build_multi_target(self):
         meta = self.meta
@@ -223,7 +223,7 @@ class TestMetadata(unittest.TestCase):
 
         # If listBuilds returns nothing, no build should be returned
         builds = []
-        self.assertIsNone(meta.get_latest_build(default=None))
+        self.assertIsNone(meta.get_latest_brew_build(default=None))
 
         meta.meta_type = 'rpm'
 
@@ -232,22 +232,22 @@ class TestMetadata(unittest.TestCase):
             self.build_record(now, assembly='not_ours', is_rpm=True, el_target=8),
             self.build_record(now, assembly='stream', is_rpm=True, el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])
 
         builds = [
             self.build_record(now, assembly='not_ours', is_rpm=True, el_target=8),
             self.build_record(now, assembly='stream', is_rpm=True, el_target=8),
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])  # No target should find el7 or el8
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])  # No target should find el7 or el8
 
         builds = [
             self.build_record(now, assembly='not_ours', is_rpm=True, el_target=8),
             self.build_record(now, assembly='stream', is_rpm=True, el_target=7),
             self.build_record(now - datetime.timedelta(hours=1), assembly='stream', is_rpm=True, el_target=8)
         ]
-        self.assertEqual(meta.get_latest_build(default=None), builds[1])  # Latest is el7 by one hour
-        self.assertEqual(meta.get_latest_build(default=None, el_target=7), builds[1])
-        self.assertEqual(meta.get_latest_build(default=None, el_target=8), builds[2])
+        self.assertEqual(meta.get_latest_brew_build(default=None), builds[1])  # Latest is el7 by one hour
+        self.assertEqual(meta.get_latest_brew_build(default=None, el_target=7), builds[1])
+        self.assertEqual(meta.get_latest_brew_build(default=None, el_target=8), builds[2])
 
 
 if __name__ == '__main__':
